+++ home-manager/modules/tridactyl.nix
@@ -18,7 +18,7 @@ in
 
     bepo = mkOption {
       type = types.bool;
-      default = false;
+      default = config.confkit.keyboard.bepo;
       example = true;
       description = "Use keybindings optimised for BÉPO keyboards.";
     };